+++ src/vector.c
@@ -413,8 +413,8 @@ vector_send(PG_FUNCTION_ARGS)
 	StringInfoData buf;
 
 	pq_begintypsend(&buf);
-	pq_sendint16(&buf, vec->dim);
-	pq_sendint16(&buf, vec->unused);
+	pq_sendint16(&buf, (uint16) vec->dim);
+	pq_sendint16(&buf, (uint16) vec->unused);
 	for (int i = 0; i < vec->dim; i++)
 		pq_sendfloat4(&buf, vec->x[i]);
